pyproject.toml
[project]
name = "vibe-trade-execution"
version = "0.1.0"
description = "Execution engine for trading strategies based on schema definitions"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "jsonschema>=4.20.0",
    "google-cloud-pubsub>=2.18.0",
    "vibe-trade-shared @ git+https://github.com/davidtoddbentz/vibe-trade-shared.git@main",
]
authors = [
    {name = "Vibe Trade", email = "dev@vibe-trade.com"},
]
license = {text = "MIT"}
keywords = ["trading", "lean", "execution", "strategy", "quantconnect"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

